fix:fix clean.sh (#2)

+++ script/clean.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
 
-docker rm `docker ps -a -q`
+if [[ "`docker ps -a -q`" != "" ]]; then
+    docker rm `docker ps -a -q`
+fi
 
 if [[ "`docker volume ls -q`" == *r0uei_db-store* ]]; then
-    docker volume rm r0uei_db-store
+    docker volume rm `docker volume ls -q`
 fi
 
 echo "0" > ./app/config/.r0uei_database_exist
